Refactor exception handling by introducing ProcessingError and updating imports across modules
