+++ src/forklift/engine/__init__.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
-from .forklift_core import ForkliftCore, ImportConfig, ProcessingResults, HeaderMode
+from .forklift_core import ForkliftCore
+from .config import ImportConfig, ProcessingResults, HeaderMode
+from .exceptions import ProcessingError
 from typing import Any, Optional
 
 # Export the main classes for backwards compatibility
@@ -7,5 +9,6 @@ __all__ = [
     "ForkliftCore",
     "ImportConfig",
     "ProcessingResults",
-    "HeaderMode"
+    "HeaderMode",
+    "ProcessingError"
 ]
