+++ src/forklift/engine/exceptions.py
@@ -1,0 +1,6 @@
+"""Exception classes for Forklift engine."""
+
+
+class ProcessingError(Exception):
+    """Raised when data processing fails."""
+    pass
